refactor(client): More refactoring
